Rewrite event loops to offer a processEvents-type interface.
Instead of always owning the main loop, we now offer up a way to "crank" the
event loop by hand. This is useful for embedding-type scenarios, where we may
want to stick rengine inside something else.

As part of this, virtual run() is now nonvirtual, and replaced with a virtual
processEvents.