Refactor Surface / SurfaceInterface and backend.
This only fixes the SDL backend. Qt backend and SFHWC backends are left
out in the cold for now.

We're limiting ourselves to a single window in the implementation, but
the API doesn't enforce this so we're free to change it down the line.

However, the limiting makes our life very easy in the backends, as
we only have to deal with a single implementation class there, the
backend itself.