input: Another iteration... demo now has button hover and click :)
The nodes have to be able to decide for themselves what to do
with a pointer event, so change the implementation so we deliver
the event to Node::onPointerEvent() after verifying it as a valid
target.

This approach is a bit nicer as we don't have to reimplement quite
as much in our own window impl.

As the demo uses touch, delete the old touch example.

There is still a bit of logic missing for this to work flawlessly:
 - Since the button sets itself as the pointer target to handle
   hover-leave, it also ends up breaking paint-order target
   selection in the case of overlapping elements.
 - The pointer target concept implies a single target, multitouch
   would allow for multiple...
We'll get back to these at a later time..